fix(runtime): version-check cached deps; consolidate satisfiesRange helper
Two items from Copilot's review of df58573:

- src/runtime/loader.ts: `ensureRuntimeInstalled` previously skipped
  any package that resolved from the cache, regardless of its
  installed version. An older doc-detective install could leave
  webdriverio/appium/sharp/etc in a persistent
  DOC_DETECTIVE_CACHE_DIR (containers, CI), and a subsequent
  doc-detective upgrade would silently run against the stale cache.
  Now the skip-when-cached path also runs the installed version
  through `satisfiesRange(installed, getDeclaredVersion(name))` —
  if the cached version doesn't satisfy the current package.json
  constraint, the package is reinstalled.
- src/runtime/heavyDeps.ts: `satisfiesRange` (and its
  `parseSemverCore` / `compareTuple` helpers) moved from
  installer.ts to heavyDeps.ts so both the installer's `status()`
  and the loader's skip-check share the same range semantics.
  installer.ts now imports it.
- src/runtime/installer.ts: the orphaned JSDoc block left behind
  by the previous round (when `satisfiesRange` lived inline)
  collapsed into a single block on `status()` with a note
  explaining the helper now lives in heavyDeps.
